# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Lorem Ipsum",
        "contact": {
            "email": "hi.wasissubekti02@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/contact-us": {
            "get": {
                "tags": [
                    "Contact Us"
                ],
                "summary": "ambil semua contact us",
                "description": "Mengambil Data Semua contact us",
                "operationId": "contact_us",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "success get all contact_us",
                                    "data": [
                                        {
                                            "id": 5,
                                            "name": "ja3hyun okkk",
                                            "email": "leepresent@tes.com",
                                            "phone": "081231231",
                                            "address": "qt",
                                            "message": "watch it dabom okk",
                                            "created_at": "2024-06-07T05:00:21.000000Z",
                                            "updated_at": "2024-06-11T08:20:45.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contact Us"
                ],
                "summary": "Create new contact us",
                "description": "Create a new contact us",
                "operationId": "create_contact_us",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of contact us",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "The email of contact us",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "The phone of contact us",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "The address of contact us",
                                        "type": "string"
                                    },
                                    "message": {
                                        "description": "The messae of contact us",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success post contact us",
                                    "data": {
                                        "name": "tbzuyeon",
                                        "email": "tbzuyeonie@tes.com",
                                        "phone": "081231231",
                                        "address": "lorem ipsumieeez",
                                        "message": "watch it dabom",
                                        "created_at": "2024-06-14T02:31:17.000000Z",
                                        "updated_at": "2024-06-14T02:31:17.000000Z",
                                        "id": 9
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail post contact us",
                                    "error": "Validation error message"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contact-us/{id}": {
            "get": {
                "tags": [
                    "Contact Us"
                ],
                "summary": "Get single contact us",
                "description": "Retrieve a single contact us by ID",
                "operationId": "single_contact_us",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the contact us",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success get contact us",
                                    "data": {
                                        "id": 5,
                                        "name": "ja3hyun okkk",
                                        "email": "leepresent@tes.com",
                                        "phone": "081231231",
                                        "address": "qt",
                                        "message": "watch it dabom okk",
                                        "created_at": "2024-06-07T05:00:21.000000Z",
                                        "updated_at": "2024-06-11T08:20:45.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "contact us not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail get contact us",
                                    "error": "No query results for model [App\\\\Models\\\\contactUs] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contact Us"
                ],
                "summary": "Update contact us",
                "description": "Update contact us by ID",
                "operationId": "update_contact_us",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of contact us",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "The description of contact us",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "The image of contact us",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success update contact us",
                                    "data": {
                                        "id": 5,
                                        "name": "ja3hyun okkk",
                                        "email": "leepresent@tes.com",
                                        "phone": "081231231",
                                        "address": "qt",
                                        "message": "watch it dabom okk",
                                        "created_at": "2024-06-07T05:00:21.000000Z",
                                        "updated_at": "2024-06-11T08:20:45.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "contact us not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail update contact us",
                                    "error": "No query results for model [App\\\\Models\\\\contactUs] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contact Us"
                ],
                "summary": "Delete contact us",
                "description": "Delete contact us by ID",
                "operationId": "delete_contact_us",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the contact us",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success delete contact us"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "contact us not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail delete contact us",
                                    "error": "No query results for model [App\\\\Models\\\\contactUs] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}